+++ config/index.js
@@ -17,6 +17,8 @@ config.title = "Alex D.'s Nuts";
 config.siteTitle = "AlexDeezNuts.com";
 config.domain = "alexdeeznuts.com";
 if (config.local) config.domain = "localhost";
+if (config.debugging) config.domain = "http://"+config.domain;
+else config.domain = "https://"+config.domain;
 config.author = "Skeetzo";
 config.description = "Porn Star Streamer";
 config.Google_Analytics = "UA-82463743-8";
@@ -31,13 +33,16 @@ config.bitcoin_address = "7h15157o74lly4b17co1n4ddre55";
 config.bitcoin_qr = "http://placehold.it/150x150";
 config.bitcoin_link = "bitcoin:"+config.bitcoin_address;
 config.blockchainRoute = "/tip";
-config.blockchainCallback = 'https://'+config.domain+config.blockchainRoute;
+config.blockchainCallback = config.domain+config.blockchainRoute;
 config.blockchainConfirmationLimit = 6;
 config.blockchainGapLimit = 20;
 config.blockchainSecret = "gofuckyourself6969";
 // RTMP Stream
 config.streamKey = "yourmotherisadirtywhore";
 config.streamKeyExpire = 3600000;
+
+var live_url = config.domain+":8443/live/stream.flv?sign=";
+if (config.debugging) live_url = config.domain+":8000/live/stream.flv";
 
 config.siteData = 
 	{ 	
@@ -47,7 +52,8 @@ config.siteData =
 		author: config.author,
 		description: config.description,
 		Google_Analytics: config.Google_Analytics,
-		status: config.status
+		status: config.status,
+		live_url: live_url
 	};
 
 config.alexd = {