+++ modules/rtmp.js
@@ -3,34 +3,19 @@ var config = require('../config/index'),
 
 const NodeMediaServer = require('node-media-server');
 
-const serverOptions = {
-  logType: 3,
+var serverOptions = {
+  'logType': 3,
 
-  rtmp: {
-    port: 1935,
-    chunk_size: 60000,
-    gop_cache: true,
-    ping: 60,
-    ping_timeout: 30
+  'rtmp': {
+    'port': 1935,
+    'chunk_size': 60000,
+    'gop_cache': true,
+    'ping': 60,
+    'ping_timeout': 30
   },
-  http: {
-    port: 8000,
-    allow_origin: '*'
-  },
-  https: {
-    port: 8443,
-    key:'./dev/privatekey.pem',
-    cert:'./dev/certificate.pem',
-  },
-  auth: {
-    play: true,
-    publish: true,
-    secret: config.streamKey
-  },
-  auth: {
-    api : true,
-    api_user: 'admin',
-    api_pass: 'rtmpsucksdeeck',
+  'http': {
+    'port': 8000,
+    'allow_origin': '*'
   },
   // trans: {
   //   ffmpeg: '/usr/local/bin/ffmpeg',
@@ -44,6 +29,19 @@ const serverOptions = {
   //   ]
   // }
 };
+
+if (!config.debugging) {
+  serverOptions.https = {
+    'port': 8443,
+    'key':'./dev/privatekey.pem',
+    'cert':'./dev/certificate.pem'
+  };
+  serverOptions.auth = {
+    'api' : true,
+    'api_user': 'admin',
+    'api_pass': 'rtmpsucksdeeck'
+  };
+}
 
 var nms = new NodeMediaServer(serverOptions);
 nms.run();